refactor: split gameplay package into gameplay, level, maze, and entities; move MazeCollisionManager to managers
