+++ core/src/main/java/com/p1_7/game/level/RoomAssigner.java
@@ -1,4 +1,6 @@
-package com.p1_7.game.gameplay;
+package com.p1_7.game.level;
+
+import com.p1_7.game.gameplay.MathQuestion;
 
 import java.util.HashMap;
 import java.util.List;
