+++ core/src/main/java/com/p1_7/game/scenes/GameScene.java
@@ -18,13 +18,13 @@ import com.p1_7.abstractengine.transform.ITransform;
 import com.p1_7.game.core.Transform2D;
 import com.p1_7.game.ui.BrightnessOverlay;
 import com.p1_7.game.ui.QuestionPanel;
+import com.p1_7.game.entities.Player;
 import com.p1_7.game.gameplay.Difficulty;
-import com.p1_7.game.gameplay.ILevelOrchestrator;
-import com.p1_7.game.gameplay.MazeCollisionManager;
-import com.p1_7.game.gameplay.MazeLayout;
-import com.p1_7.game.gameplay.Player;
 import com.p1_7.game.gameplay.RoundPhase;
-import com.p1_7.game.gameplay.WallCollidable;
+import com.p1_7.game.level.ILevelOrchestrator;
+import com.p1_7.game.managers.MazeCollisionManager;
+import com.p1_7.game.maze.MazeLayout;
+import com.p1_7.game.maze.WallCollidable;
 import com.p1_7.game.managers.GameMovementManager;
 import com.p1_7.game.managers.IFontManager;
 import com.p1_7.game.platform.GdxDrawContext;
